+++ src/OpenLoco/src/Ui/Windows/ObjectSelectionWindow.cpp
@@ -497,9 +497,10 @@ namespace OpenLoco::Ui::Windows::ObjectSelectionWindow
     }
 
     static const WindowEventList& getEvents();
+    static void switchPrimaryTab(Window& self, uint8_t tabIndex);
+    static void switchTabByObjectType(Window& self, ObjectType objectType);
 
-    // 0x00472A20
-    Ui::Window* open()
+    static Ui::Window* internalOpen(std::optional<ObjectType> optionalObjectType)
     {
         auto window = WindowManager::bringToFront(WindowType::objectSelection);
 
@@ -541,18 +542,37 @@ namespace OpenLoco::Ui::Windows::ObjectSelectionWindow
         inputSession = Ui::TextInput::InputSession();
         inputSession.calculateTextOffset(widgets[widx::textInput].width());
 
+        if (optionalObjectType.has_value())
+        {
+            ObjectType objectType = optionalObjectType.value();
+
+            window->filterLevel = enumValue(FilterLevel::advanced);
+            assignTabPositions(window);
+            switchTabByObjectType(*window, objectType);
+        }
+
+        // If in play mode, the object selection window should be modal and pause the game to prevent issues.
+        if (!SceneManager::isEditorMode())
+        {
+            WindowManager::setCurrentModalType(WindowType::objectSelection);
+
+            SceneManager::setPauseFlag(1 << 3); // Pause flag 3 was not used by vanilla.
+            WindowManager::invalidate(WindowType::timeToolbar);
+            Audio::pauseSound();
+        }
+
         return window;
     }
 
-    static void switchPrimaryTab(Window& self, uint8_t tabIndex);
-    static void switchTabByObjectType(Window& self, ObjectType objectType);
+    // 0x00472A20
+    Ui::Window* open()
+    {
+        return internalOpen({});
+    }
 
     Window& openInTab(ObjectType objectType)
     {
-        auto& window = *open();
-        window.filterLevel = enumValue(FilterLevel::advanced);
-        assignTabPositions(&window);
-        switchTabByObjectType(window, objectType);
+        auto& window = *internalOpen(objectType);
         return window;
     }
 
@@ -1591,6 +1611,13 @@ namespace OpenLoco::Ui::Windows::ObjectSelectionWindow
             Gfx::invalidateScreen();
             CompanyManager::determineAvailableVehicles();
             WindowManager::invalidate(WindowType::buildVehicle);
+
+            // Stop being modal and unpause game.
+            WindowManager::setCurrentModalType(WindowType::undefined);
+
+            SceneManager::unsetPauseFlag(1 << 3);
+            WindowManager::invalidate(WindowType::timeToolbar);
+            Audio::unpauseSound();
         }
         ObjectManager::freeSelectionList();
     }